HtmlBridge Phase 3 ratchet: de-list SubDocuments.cs (split fragment-mutation + XML partials)
SubDocuments.cs (1152 lines) carried two concerns beyond its name; both split
into sibling partials:

- DomBridge/HtmlFragmentMutation.cs (329 lines): the generic HTML-fragment
  DOM-mutation helpers — RemoveElementsRecursive, NormalizeNode, RemoveChildAt,
  and the innerHTML/outerHTML/insertAdjacentHTML/child-node-argument fragment
  builders (SetElementInnerHtml/OuterHtml, BuildAdjacentHtmlNodes,
  BuildChildNodeArgumentNodes, InsertNodeAt, GetInsertAdjacentTarget,
  TryBuildInnerHtmlFragmentContainer, IsVoidHtmlElementTag).
- DomBridge/SubDocuments.XmlAndScripts.cs (163 lines): XML/XHTML sub-document
  construction and sub-document script execution (BuildSubDocumentFromXml,
  BuildDomElementFromXElement, ExecuteSubDocumentScripts, CollectScriptContent,
  GetTextContentRecursive).

A dead orphan doc-comment at the file tail was dropped, and the now-unused
System.Text / System.Xml.Linq usings were removed from the original.

Pure partial-class relocation — no signature, accessibility, or logic change,
so behaviour-identical by construction. SubDocuments.cs drops 1152 -> 694,
under the 750-line guard, so its OversizedFileExemptions entry is removed and
the stale-exemption check enforces it stays under the limit. Debt list:
five -> four files.

Broiler.HtmlBridge.Dom builds clean; HtmlBridgeArchitectureGuardTests 14/14
green (0 offenders / 0 stale exemptions); fragment-mutation / sub-document
coverage (InnerHtmlParallelStateRemovalTests, SubDocumentBindingModuleTests,
SubDocumentSeverMigrationTests, DocumentFragmentSentinelMigrationTests,
DomImplementationTests, HtmlDomInterfacesTests, DomEdgeCasePhase4Tests,
IsEqualNodePromotionTests, SvgDomAndCrossDocTests, Acid3RegressionTests,
204 tests with the guard) green.
